upd commands partly
commands was partly updated for new api
