--- Commands/DataBase/GetAccountCommand.php
@@ -1,16 +1,0 @@
-<?php
-
-
-namespace GrapheneNodeClient\Commands\DataBase;
-
-
-class GetAccountCommand extends CommandAbstract
-{
-    /** @var string */
-    protected $method = 'get_accounts';
-
-    /** @var array */
-    protected $queryDataMap = [
-        '0' => ['array'], //authors
-    ];
-}
