+++ Commands/Single/GetAccountsCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+class GetAccountsCommand extends CommandAbstract
+{
+    /** @var string */
+    protected $method = 'get_accounts';
+}
