+++ Commands/Single/GetDiscussionsByCreatedCommand.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace GrapheneNodeClient\Commands\Single;
+
+
+
+class GetDiscussionsByCreatedCommand extends CommandAbstract
+{
+    protected $method       = 'get_discussions_by_created';
+}
